pkg/http/middleware/cors.go
package middleware

import (
	"github.com/gin-gonic/gin"
	"net/http"
	"strings"
)

var allowHeaders = []string{
	"Content-Type",
	"Content-Length",
	"Accept-Encoding",
	"X-CSRF-Token",
	"Authorization",
	"Origin",
	"Cache-Control",
	"X-Requested-With",
	"Cookie",
}

var allowOrigins = []string{
	"http://localhost:3000",
}

func CORS() gin.HandlerFunc {
	return func(ctx *gin.Context) {
		ctx.Writer.Header().Set("Access-Control-Allow-Origin", strings.Join(allowOrigins, ","))
		ctx.Writer.Header().Set("Access-Control-Allow-Credentials", "true")
		ctx.Writer.Header().Set("Access-Control-Allow-Headers", strings.Join(allowHeaders, ","))
		ctx.Writer.Header().Set("Access-Control-Allow-Methods", "GET, POST, PATCH, DELETE")

		if ctx.Request.Method == "OPTIONS" {
			ctx.AbortWithStatus(http.StatusNoContent)
			return
		}

		ctx.Next()
	}
}
